+++ examples/shared_samples/server_hardware_type.rb
@@ -14,7 +14,7 @@ require_relative '../_client' # Gives access to @client
 # Example: Actions with a Server Hardware Type
 #
 # Supported APIs:
-# - 200, 300, 500
+# - 200, 300, 500, 600
 
 # Resources that can be created according to parameters:
 # api_version = 200 & variant = any to OneviewSDK::API200::ServerHardwareType
@@ -22,6 +22,8 @@ require_relative '../_client' # Gives access to @client
 # api_version = 300 & variant = Synergy to OneviewSDK::API300::Synergy::ServerHardwareType
 # api_version = 500 & variant = C7000 to OneviewSDK::API500::C7000::ServerHardwareType
 # api_version = 500 & variant = Synergy to OneviewSDK::API500::Synergy::ServerHardwareType
+# api_version = 600 & variant = C7000 to OneviewSDK::API600::C7000::ServerHardwareType
+# api_version = 600 & variant = Synergy to OneviewSDK::API600::Synergy::ServerHardwareType
 
 # Resource Class used in this sample
 shw_type_class = OneviewSDK.resource_named('ServerHardwareType', @client.api_version)