+++ lib/oneview-sdk/resource/api600/c7000/logical_switch_group.rb
@@ -9,13 +9,23 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-require_relative '../../api300/c7000/logical_switch_group'
+require_relative '../../api500/c7000/logical_switch_group'
 
 module OneviewSDK
   module API600
     module C7000
       # Logical switch group resource implementation for API600 C7000
-      class LogicalSwitchGroup < OneviewSDK::API300::C7000::LogicalSwitchGroup
+      class LogicalSwitchGroup < OneviewSDK::API500::C7000::LogicalSwitchGroup
+        # Create a resource object, associate it with a client, and set its properties.
+        # @param [OneviewSDK::Client] client The client object for the OneView appliance
+        # @param [Hash] params The options for this resource (key-value pairs)
+        # @param [Integer] api_ver The api version to use when interracting with this resource.
+        def initialize(client, params = {}, api_ver = nil)
+          @data ||= {}
+          # Default values:
+          @data['type'] ||= 'logical-switch-groupV4'
+          super
+        end
       end
     end
   end