+++ spec/integration/resource/api600/c7000/connection_template/update_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::ConnectionTemplate
+klass = OneviewSDK::API600::C7000::ConnectionTemplate
 RSpec.describe klass, integration: true, type: UPDATE do
-  let(:current_client) { $client_500 }
-  include_examples 'ConnectionTemplateUpdateExample', 'integration api500 context'
+  let(:current_client) { $client_600 }
+  include_examples 'ConnectionTemplateUpdateExample', 'integration api600 context'
 end