+++ spec/integration/resource/api600/c7000/logical_switch/create_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::LogicalSwitch
+klass = OneviewSDK::API600::C7000::LogicalSwitch
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500 }
-  let(:logical_switch_group_class) { OneviewSDK::API500::C7000::LogicalSwitchGroup }
-  include_examples 'LogicalSwitchCreateExample', 'integration api500 context'
+  let(:current_client) { $client_600 }
+  let(:logical_switch_group_class) { OneviewSDK::API600::C7000::LogicalSwitchGroup }
+  include_examples 'LogicalSwitchCreateExample', 'integration api600 context'
 end