+++ spec/integration/resource/api600/c7000/uplink_set/create_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::UplinkSet
+klass = OneviewSDK::API600::C7000::UplinkSet
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:li_name) { LOG_INT_NAME }
-  include_examples 'UplinkSetCreateExample', 'integration api500 context'
+  include_examples 'UplinkSetCreateExample', 'integration api600 context'
 end