+++ spec/integration/resource/api600/synergy/volume_template/create_spec.rb
@@ -11,11 +11,11 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::VolumeTemplate
+klass = OneviewSDK::API600::Synergy::VolumeTemplate
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500_synergy }
+  let(:current_client) { $client_600_synergy }
   let(:storage_system_ip) { $secrets_synergy['storage_system1_ip'] }
   let(:storage_virtual_ip) { $secrets_synergy['store_virtual_ip'] }
-  include_context 'integration api500 context'
-  include_examples 'VolumeTemplateCreateExample API500'
+  include_context 'integration api600 context'
+  include_examples 'VolumeTemplateCreateExample API600'
 end